Extract DialogFields mixin
