+++ app/controllers/api/base_controller/generic.rb
@@ -142,16 +142,6 @@ module Api
         invoke_custom_action(type, resource, action, data)
       end
 
-      def refresh_dialog_fields_action(dialog, refresh_fields, resource_ident)
-        result = {}
-        refresh_fields.each do |field|
-          dynamic_field = dialog.field(field)
-          return action_result(false, "Unknown dialog field #{field} specified") unless dynamic_field
-          result[field] = dynamic_field.update_and_serialize_values
-        end
-        action_result(true, "Refreshing dialog fields for #{resource_ident}", :result => result)
-      end
-
       private
 
       def add_subcollection_data_to_resource(resource, type, subcollection_data)
