+++ app/controllers/api/service_catalogs_controller.rb
@@ -1,5 +1,6 @@
 module Api
   class ServiceCatalogsController < BaseController
+    include Shared::DialogFields
     include Subcollections::ServiceTemplates
   end
 end
