+++ app/controllers/api/service_templates_controller.rb
@@ -1,5 +1,6 @@
 module Api
   class ServiceTemplatesController < BaseController
+    include Shared::DialogFields
     include Subcollections::ServiceDialogs
     include Subcollections::Tags
     include Subcollections::ResourceActions
